[Experimental] Back to use svcSleepThread to delay
Back to commit 44dc73c I did not understand why the delay method from
CtrBootManager (basically waiting for frames) worked better tham my
method using svcSleepThread (the "proper way", equivalent to usleep()).
I thought it was because it flushed buffers, and since CakeBrah is all
about GFX this made sense.

However now I know why. Since the other method waited for frames, and
3DS uses 60FPS, with the other method setting delay to 100 means that
you wait for 100 / 60 (FPS) = 1.6667s, while 100 with my method means,
literally 100ms or 0.1s. So yeah, I missed the delay time by a lot.

Returning back to the older method. It is way more simple (1 function
call instead of 3 and a for loop), makes much more sense and don't
need people to make math just to get the desired delay. However set a
proper default delay this time (1500, or 1.5ms). I could use 1667 too,
however this seems to work (and I already changed the default delay to
90 in the previous commit, since 90 / 60 (FPS) = 1.5s).

Update the docs to reflect the delay changes too.

*: btw, this is experimental because I am not sure that doing this will
be equivalent to the old method. However it seems to work well in my
tests.

+++ source/config.h
@@ -10,7 +10,7 @@ typedef struct {
     // "path" and "delay" are the results read from handler() function
     char *key;
     char *path;
-    unsigned int delay;
+    unsigned long long delay;
     int payload;
     unsigned long offset;
 } configuration;
